requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.8.0
dnspython==2.7.0
docutils==0.21.2
Flask==3.1.0
Flask-PyMongo==3.0.1
Flask-WTF==1.2.2
flit==3.12.0
flit_core==3.12.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pydantic==2.11.4
pydantic_core==2.33.2
pymongo==4.12.1
pytest==8.3.5
requests==2.32.3
tomli_w==1.2.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
